Refactoring. Added click-listener
